Update zteleport.mcfunction
Added support for the new world height, fixed an issue with one of the lines reading "230" instead of "250" in blockcount. Added support for 10000 and 15000. These numbers are not perfectly accurate and may be slightly smaller than the actual resulting size. Added a little extra leeway in case ov overteleportation, this should not happen, but it is a precaution.